Улучшил обработку и вывод ошибок в сервисе закзов

+++ orders_service/src/main/java/com/example/orders_service/service/implementation/WarehouseImpl.java
@@ -3,19 +3,28 @@ package com.example.orders_service.service.implementation;
 import com.example.orders_service.domain.DishException;
 import com.example.orders_service.dto.DishDto;
 import com.example.orders_service.dto.DishInfoForManagerDto;
+import com.example.orders_service.dto.DishInfoForUserDto;
 import com.example.orders_service.entity.Dish;
 import com.example.orders_service.repository.DishRepository;
 import com.example.orders_service.service.abstraction.DishMapper;
-import com.example.orders_service.service.abstraction.DishService;
+import com.example.orders_service.service.abstraction.Warehouse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
-public class DishServiceImpl implements DishService {
+public class WarehouseImpl implements Warehouse {
     private final DishRepository dishRepository;
 
     private final DishMapper dishMapper;
+
+    @Override
+    public List<DishInfoForUserDto> getAllAvailableDishes() {
+        return dishRepository.getAllWhereIsAvailableAndQuantityMoreThanZero()
+                .stream().map(dishMapper::mapToDishForUserDto).toList();
+    }
 
     @Override
     public long createNewDish(DishDto dishDto) {
@@ -27,14 +36,14 @@ public class DishServiceImpl implements DishService {
     @Override
     public DishInfoForManagerDto getDishById(long id) {
         Dish dish = dishRepository.findById(id)
-                .orElseThrow(() -> new DishException("Блюда с таким id не существует"));
+                .orElseThrow(DishException::dishNotFound);
         return dishMapper.mapToDishForManagerDto(dish);
     }
 
     @Override
     public void updateDishById(long id, DishDto dishDto) {
-        if (dishRepository.findById(id).isEmpty()) {
-            throw new DishException("Блюда с таким id не существует");
+        if (!dishRepository.existsById(id)) {
+            throw DishException.dishNotFound();
         }
         Dish dish = dishMapper.mapToDishEntity(dishDto);
         dish.setId(id);
@@ -43,6 +52,9 @@ public class DishServiceImpl implements DishService {
 
     @Override
     public void deleteDishById(long id) {
+        if (!dishRepository.existsById(id)) {
+            throw DishException.dishNotFound();
+        }
         dishRepository.deleteById(id);
     }
 }
